requirements.txt
 openai>=1.40.0,<2.0.0
 python-dotenv>=1.0.1,<2.0.0

faster-whisper>=1.0.0,<2.0.0
sounddevice>=0.4.7,<0.5.0
soundfile>=0.12.1,<0.13.0

